create control folder
