+++ nanome/api/callbacks.py
@@ -7,14 +7,6 @@ from nanome.api.integration import Integration, IntegrationRequest
 logger = logging.getLogger(__name__)
 
 
-def advanced_settings(network, args, request_id):
-    network.on_advanced_settings()
-
-
-def connect(network, arg, request_id):
-    pass
-
-
 def integration(network, args, request_id):
     integration = network._plugin.integration
     request = IntegrationRequest(args[0], args[1], args[2], network)
