+++ nanome/api/control/callbacks.py
@@ -1,0 +1,11 @@
+
+def advanced_settings(network, args, request_id):
+    network.on_advanced_settings()
+
+
+def connect(network, arg, request_id):
+    pass
+
+
+def run(network, args, request_id):
+    network._on_run()
